restructuré la manière de sauver une Map. doit encore rajouter le chargement d'une Map avec la même technique, faire fonctionner les thumbnails avec, et créer le même système pour le jeu
